fix(core): move prepareFlattenedInnerRoot to inlineSubCompositions to avoid runtime-inline dependency
The test import of prepareFlattenedInnerRoot from htmlBundler pulled in
the generated/runtime-inline module which doesn't exist in CI test runs
(only after build). Moving the function to inlineSubCompositions breaks
the transitive dependency chain while keeping the re-export from
htmlBundler for backwards compatibility.
