+++ packages/core/src/compiler/inlineSubCompositions.ts
@@ -371,3 +371,41 @@ export function inlineSubCompositions(
 
   return { styles, scripts, externalScriptSrcs, scriptItems, externalLinks, variablesByComp };
 }
+
+// ---------------------------------------------------------------------------
+// Inner root flattening (shared by bundler and producer)
+// ---------------------------------------------------------------------------
+
+export const FLATTENED_INNER_ROOT_STRIP_ATTRS = [
+  "data-composition-id",
+  "data-composition-file",
+  "data-start",
+  "data-duration",
+  "data-end",
+  "data-track-index",
+  "data-track",
+  "data-composition-src",
+  "data-hf-authored-duration",
+  "data-hf-authored-end",
+];
+
+export function prepareFlattenedInnerRoot(innerRoot: Element): Element {
+  const prepared = innerRoot.cloneNode(true) as Element;
+  const authoredRootId = prepared.getAttribute("id")?.trim();
+  for (const attrName of FLATTENED_INNER_ROOT_STRIP_ATTRS) {
+    prepared.removeAttribute(attrName);
+  }
+  if (authoredRootId) {
+    prepared.removeAttribute("id");
+    prepared.setAttribute("data-hf-authored-id", authoredRootId);
+  }
+  prepared.setAttribute("data-hf-inner-root", "true");
+  const w = prepared.getAttribute("data-width");
+  const h = prepared.getAttribute("data-height");
+  const widthVal = w ? `${w}px` : "100%";
+  const heightVal = h ? `${h}px` : "100%";
+  const existingStyle = (prepared.getAttribute("style") || "").trim();
+  const fill = `width:${widthVal};height:${heightVal}`;
+  prepared.setAttribute("style", existingStyle ? `${existingStyle};${fill}` : fill);
+  return prepared;
+}
